Fix warning pt1

+++ Project/Engine/src/Graphics/DebugDraw/DebugDrawSystem.cpp
@@ -95,7 +95,7 @@ void DebugDrawSystem::CreateCubeGeometry()
     // Position attribute
     cubeGeometry.vao->LinkAttrib(*cubeGeometry.vbo, 0, 3, GL_FLOAT, sizeof(glm::vec3), (void*)0);
 
-    cubeGeometry.indexCount = indices.size();
+    cubeGeometry.indexCount = static_cast<unsigned int>(indices.size());
     cubeGeometry.vao->Unbind();
 }
 
@@ -191,7 +191,7 @@ void DebugDrawSystem::CreateSphereGeometry()
 
     sphereGeometry.vao->LinkAttrib(*sphereGeometry.vbo, 0, 3, GL_FLOAT, sizeof(glm::vec3), (void*)0);
 
-    sphereGeometry.indexCount = indices.size();
+    sphereGeometry.indexCount = static_cast<unsigned int>(indices.size());
     sphereGeometry.vao->Unbind();
 }
